unshare: move some code from cmd/buildah/unshare
so that it can be reused by other programs.
