+++ unshare/unshare_cgo.go
@@ -3,8 +3,8 @@
 package unshare
 
 // #cgo CFLAGS: -Wall
-// extern void _buildah_unshare(void);
+// extern void _containers_unshare(void);
 // void __attribute__((constructor)) init(void) {
-//   _buildah_unshare();
+//   _containers_unshare();
 // }
 import "C"
